+++ pubspec.yaml
@@ -14,6 +14,7 @@ dependencies:
   firebase_core: ^1.13.1
   flutter:
     sdk: flutter
+  flutter_hooks: ^0.18.4
   roggle: ^0.2.1+1
 
 dev_dependencies:
